requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
PyMySQL==1.0.2
python-dotenv==0.21.0
Werkzeug==3.0.3
email-validator==1.3.1
qrcode==7.4.2
Pillow==10.1.0
